requirements.txt
python-telegram-bot==21.0.1
Pillow==10.3.0
